Add basic backend authentication

+++ src/main/java/ch/zli/m223/punchclock/controller/EntryController.java
@@ -2,12 +2,13 @@ package ch.zli.m223.punchclock.controller;
 
 import java.util.List;
 
+import javax.annotation.security.RolesAllowed;
 import javax.inject.Inject;
 import javax.validation.Valid;
-import javax.validation.Validator;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.NotAuthorizedException;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -19,7 +20,10 @@ import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 
 import ch.zli.m223.punchclock.domain.Entry;
+import ch.zli.m223.punchclock.domain.User;
 import ch.zli.m223.punchclock.service.EntryService;
+import ch.zli.m223.punchclock.service.UserService;
+import io.quarkus.security.runtime.SecurityIdentityAssociation;
 
 @Path("/entries")
 @Tag(name = "Entries", description = "Handling of entries")
@@ -28,35 +32,63 @@ public class EntryController {
     @Inject
     EntryService entryService;
 
+    @Inject
+    UserService userService;
+
+    @Inject
+    SecurityIdentityAssociation identity;
+
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    @Operation(summary = "List all Entries", description = "")
+    @RolesAllowed({"User", "Admin"})
+    @Operation(summary = "List all Entries of the current logged in User", description = "")
     public List<Entry> list() {
-        return entryService.findAll();
+        User loggedInUser = userService.getUserByUsername(identity.getIdentity().getPrincipal().getName());
+        return loggedInUser.getEntries();
     }
 
     @POST
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
+    @RolesAllowed({"User", "Admin"})
     @Operation(summary = "Add a new Entry", description = "The newly created entry is returned. The id may not be passed.")
     public Entry add(@Valid Entry entry) throws Exception {
+        User loggedInUser = userService.getUserByUsername(identity.getIdentity().getPrincipal().getName());
+        entry.setUser(loggedInUser);
         return entryService.createEntry(entry);
     }
 
     @PUT
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
+    @RolesAllowed({"User", "Admin"})
     @Operation(summary = "Updates an entry", description = "The updated entry is returned")
     public Entry update(@Valid Entry entry) {
-        return entryService.updateEntry(entry);
+        User loggedInUser = userService.getUserByUsername(identity.getIdentity().getPrincipal().getName());
+        Entry foundEntry = entryService.findById(entry.getId()); //Get a copy of the entry in the database to confirm that the user has rights to edit it
+        
+        if(foundEntry.getUser().getUsername() == loggedInUser.getUsername()) {
+            return entryService.updateEntry(entry);
+        }else {
+            throw new NotAuthorizedException("User not authorized to update this entry");
+        }
+      
     }
 
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
+    @RolesAllowed({"User", "Admin"})
     @Operation(summary = "Deletes an entry", description = "The deleted entry is returned")
     @Path("/{id}")
     public Entry delete(@PathParam("id") long id) {
-        return entryService.deleteEntry(id);
+        User loggedInUser = userService.getUserByUsername(identity.getIdentity().getPrincipal().getName());
+        Entry foundEntry = entryService.findById(id); //Get a copy of the entry in the database to confirm that the user has rights to edit it
+        
+        if(foundEntry.getUser().getUsername() == loggedInUser.getUsername()) {
+            return entryService.deleteEntry(id);
+        }else {
+            throw new NotAuthorizedException("User not authorized to delete this entry");
+        }
     }
 }
